Renamed to *.php from *.php3 per recommendations for PHP4. Changes to support new database API for MySQL and PostgreSQL support. If area isn't defined; don't use 0, get a real area ID. Quote strings used as array subscripts, except within quoted strings. Round up duration to next multiple of $resolution, not 30, minutes. Don't allow 0-minute bookings; make them $resolution minutes. Recode building repeat day list to avoid PHP undefined variable warnings. HTML fixups: quote tag attributes where needed by HTML standard. Trailer now outputs end HTML tags, so remove them.
Supply repeat_id to mrbsCheckFree so it doesn't have to do a query for each
requested slot in the series.

Acquire and hold a mutex database lock while checking that the requested
slots are free and then booking them, to prevent double-booking from races.

Always set "$all" flag to mrbsDelEntry() to have it delete modified entries
in the series too. This fixes one of the "duplicate entry after
modification" bugs.


git-svn-id: http://svn.code.sf.net/p/mrbs/code/mrbs/branches/mrbs-0-9-1xp@187 25f7e783-145e-4f2d-a9f8-dd7182bfe9c7